better image comparison utility

+++ test/support/utilities.js
@@ -3,8 +3,10 @@ var util = require('util');
 var path = require('path');
 var spawn = require('child_process').spawn;
 var exec = require('child_process').exec;
+var existsSync = require('fs').existsSync || require('path').existsSync
 
 var image_magick_available = true;
+var overwrite = false;
 
 exec('compare -h', function(error, stdout, stderr) {
     if (error !== null) {
@@ -13,11 +15,20 @@ exec('compare -h', function(error, stdout, stderr) {
 });
 
 exports.imageEqualsFile = function(buffer, file, callback) {
+    file = path.resolve(file);
+    if (overwrite) {
+        var err = fs.writeFileSync(file, buffer);
+        if (err) {
+            err.similarity = 0;
+        }
+        return callback(err);
+    }
     if (!image_magick_available) {
         throw new Error("imagemagick 'compare' tool is not available, please install before running tests");
     }
-    file = path.resolve(file);
     var compare = spawn('compare', ['-metric', 'PSNR', '-', file, '/dev/null' ]);
+    var type = path.extname(file);
+    var result = path.join(path.dirname(file), path.basename(file, type) + '.result' + type);
 
     var error = '';
     compare.stderr.on('data', function(data) {
@@ -25,10 +36,12 @@ exports.imageEqualsFile = function(buffer, file, callback) {
     });
     compare.on('exit', function(code, signal) {
         if (!code && error.trim() === 'inf') {
+            if (existsSync(result)) {
+                // clean up old failures
+                fs.unlinkSync(result);
+            }
             callback(null);
         } else {
-            var type = path.extname(file);
-            var result = path.join(path.dirname(file), path.basename(file, type) + '.result' + type);
             fs.writeFileSync(result, buffer);
 
             if (code) {
